# internal/api/handlers/realtime.go
package handlers

import (
	"net/http"

	"github.com/ketsuna-org/sovrabase/internal/models"
)

// RealtimeConnectionHandler establishes a realtime websocket connection
// @Summary Both Websocket connection
// @Tags Realtime
// @Param id path string true "Project ID"
// @Success 200
// @Router /project/{id}/realtime [get]
func RealtimeConnectionHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement websocket connection logic
	w.WriteHeader(http.StatusOK)
}

// ListChannelsHandler lists all realtime channels
// @Summary List Channels
// @Tags Realtime
// @Security Bearer
// @Param id path string true "Project ID"
// @Success 200
// @Router /project/{id}/realtime/channels [get]
func ListChannelsHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement list channels logic
	w.WriteHeader(http.StatusOK)
}

// CreateChannelHandler creates a new realtime channel
// @Summary Create Channel
// @Tags Realtime
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param request body models.CreateChannelRequest true "Channel creation data"
// @Success 200
// @Router /project/{id}/realtime/channels [post]
func CreateChannelHandler(w http.ResponseWriter, r *http.Request) {
	var req models.CreateChannelRequest
	_ = req
	// TODO: Implement create channel logic
	w.WriteHeader(http.StatusOK)
}

// DeleteChannelHandler deletes a realtime channel
// @Summary Delete Channel
// @Tags Realtime
// @Security Bearer
// @Param id path string true "Project ID"
// @Param channel_id path string true "Channel ID"
// @Success 200
// @Router /project/{id}/realtime/channels/{channel_id} [delete]
func DeleteChannelHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement delete channel logic
	w.WriteHeader(http.StatusOK)
}

// BroadcastMessageHandler broadcasts a message to a channel
// @Summary Broadcast Message
// @Tags Realtime
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param channel_id path string true "Channel ID"
// @Param request body models.BroadcastMessageRequest true "Broadcast message data"
// @Success 200
// @Router /project/{id}/realtime/channels/{channel_id}/broadcast [post]
func BroadcastMessageHandler(w http.ResponseWriter, r *http.Request) {
	var req models.BroadcastMessageRequest
	_ = req
	// TODO: Implement broadcast message logic
	w.WriteHeader(http.StatusOK)
}

// GetPresenceHandler gets presence information for a channel
// @Summary Get Presence
// @Tags Realtime
// @Security Bearer
// @Param id path string true "Project ID"
// @Param channel path string true "Channel Name"
// @Success 200
// @Router /project/{id}/realtime/presence/{channel} [get]
func GetPresenceHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get presence logic
	w.WriteHeader(http.StatusOK)
}

// TrackPresenceHandler tracks presence in a channel
// @Summary Track Presence
// @Tags Realtime
// @Security Bearer
// @Accept json
// @Produce json
// @Param id path string true "Project ID"
// @Param channel path string true "Channel Name"
// @Param request body models.TrackPresenceRequest true "Presence tracking data"
// @Success 200
// @Router /project/{id}/realtime/presence/{channel}/track [post]
func TrackPresenceHandler(w http.ResponseWriter, r *http.Request) {
	var req models.TrackPresenceRequest
	_ = req
	// TODO: Implement track presence logic
	w.WriteHeader(http.StatusOK)
}

// UntrackPresenceHandler untracks presence in a channel
// @Summary Untrack Presence
// @Tags Realtime
// @Security Bearer
// @Param id path string true "Project ID"
// @Param channel path string true "Channel Name"
// @Success 200
// @Router /project/{id}/realtime/presence/{channel}/untrack [delete]
func UntrackPresenceHandler(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement untrack presence logic
	w.WriteHeader(http.StatusOK)
}